move loadpkl tests to proper location
